- Provide for a plugin storage service - Remove default of saving plugin data to database
